refactor: adopt std LazyLock, and stop suppressing clippy workspace-wide
Three related pieces of modernization now that the MSRV is 1.96.

lazy_static -> std::sync::LazyLock (stable since 1.80) across all six crates
that used it, removing the dependency from every first-party manifest; it
survives only as a transitive dep of fuse-backend-rs. Mostly mechanical, with
two wrinkles: `root_tracer!` needs `&*BUILDING_RECORDER` because a
`&LazyLock<T>` will not coerce where a `&T` is wanted, and making the type
visible to clippy (lazy_static's macro had hidden it) tripped
`type_complexity` on the proxy SDK client map -- resolved by dropping its
outer `Arc`, which was never cloned and bought nothing on a `'static`.

Removed the four blanket `--allow` flags from the clippy line. A workspace-
wide allow also covers code written later, so a stale one hides real
findings. `needless_borrow`, `result_large_err` and `io_other_error` were
completely dead; `unnecessary_cast` was hiding five redundant `addr as u64`
casts in block_uffd.rs, fixed rather than re-suppressed.

Deleted 19 dead `#[allow]` attributes, found by converting every attribute
to `#[expect]` and collecting the unfulfilled ones on both macOS and Linux
under the CI feature set. Only the intersection was removed. Six of the
original 25 candidates turned out to still be needed: dead-code reachability
cascades, so an allow that looks stale while its neighbours are still
suppressing is not necessarily stale once they are gone. Linux caught that.

Also six let-chains where an outer `if let` wrapped a lone `if`. That is far
fewer sites than the nested-`if let` count suggests -- nearly every one has
an `else` whose body needs the outer binding, which a let-chain cannot
express -- and the v6 inode `store()` comments now explain that xattrs are
appended by the caller rather than claiming they are missing.

+++ service/src/block_uffd.rs
@@ -2512,7 +2512,7 @@ mod tests {
             // so we must NOT close it via cleanup_uffd_region afterward.
             let handshake = HandshakeRequest {
                 r#type: MessageType::Handshake,
-                regions: vec![VmaRegion::new(addr as u64, 4096, 0, 4096)],
+                regions: vec![VmaRegion::new(addr, 4096, 0, 4096)],
                 policy: FaultPolicy::Zerocopy,
                 enable_prefault: false,
             };
@@ -2577,7 +2577,7 @@ mod tests {
 
             // Setup uffd region - OwnedFd takes ownership of uffd_fd
             let (uffd_fd, addr, mmap_size) = setup_uffd_region(4096).unwrap();
-            let vma_regions = vec![VmaRegion::new(addr as u64, 4096, 0, 4096)];
+            let vma_regions = vec![VmaRegion::new(addr, 4096, 0, 4096)];
             let state = ConnState {
                 vma_regions,
                 policy: FaultPolicy::Zerocopy,
@@ -3046,21 +3046,21 @@ mod tests {
         compio::runtime::Runtime::new().unwrap().block_on(async {
             let (uffd_fd, addr, mmap_size) = setup_uffd_region(4096).unwrap();
 
-            let result = uffdio_wake(uffd_fd, addr as u64, 4096).await;
+            let result = uffdio_wake(uffd_fd, addr, 4096).await;
             assert!(
                 result.is_ok(),
                 "uffdio_wake on registered region failed: {:?}",
                 result.err()
             );
 
-            let zeropage_result = uffdio_zeropage(uffd_fd, addr as u64, 4096).await;
+            let zeropage_result = uffdio_zeropage(uffd_fd, addr, 4096).await;
             assert!(
                 zeropage_result.is_ok(),
                 "uffdio_zeropage failed: {:?}",
                 zeropage_result.err()
             );
 
-            let result = uffdio_wake(uffd_fd, addr as u64, 4096).await;
+            let result = uffdio_wake(uffd_fd, addr, 4096).await;
             assert!(
                 result.is_ok(),
                 "uffdio_wake after zeropage should return Ok (EEXIST): {:?}",
